Update on 22.10.2020

+++ SZMK.TeklaInteraction.Shared/Models/Drawing.cs
@@ -15,7 +15,7 @@ namespace SZMK.TeklaInteraction.Shared.Models
         public int CountMark { get; set; }
         public double SubTotalWeight { get; set; }
         public double SubTotalLenght { get; set; }
-        public int CountDetail { get; set; }
+        public long CountDetail { get; set; }
         public TypeAdd TypeAdd { get; set; }
         public Model Model { get; set; }
         public List<Detail> Details { get; set; }
